@@ -12,15 +12,15 @@
   "private": true,
   "dependencies": {
     "@angular/animations": "^17.3.0",
-    "@angular/common": "^17.2.4",
-    "@angular/compiler": "^17.2.4",
-    "@angular/core": "^17.2.4",
-    "@angular/forms": "^17.2.4",
-    "@angular/localize": "^17.2.4",
-    "@angular/platform-browser": "^17.2.4",
-    "@angular/platform-browser-dynamic": "^17.2.4",
-    "@angular/platform-server": "^17.2.4",
-    "@angular/router": "^17.2.4",
+    "@angular/common": "^17.3.0",
+    "@angular/compiler": "^17.3.0",
+    "@angular/core": "^17.3.0",
+    "@angular/forms": "^17.3.0",
+    "@angular/localize": "^17.3.0",
+    "@angular/platform-browser": "^17.3.0",
+    "@angular/platform-browser-dynamic": "^17.3.0",
+    "@angular/platform-server": "^17.3.0",
+    "@angular/router": "^17.3.0",
     "@angular/ssr": "^17.3.0",
     "@progress/kendo-angular-buttons": "15.1.0",
     "@progress/kendo-angular-common": "15.1.0",
@@ -36,7 +36,7 @@
     "@progress/kendo-angular-popup": "15.1.0",
     "@progress/kendo-angular-progressbar": "15.1.0",
     "@progress/kendo-font-icons": "^2.1.0",
-    "@progress/kendo-licensing": "^1.0.2",
+    "@progress/kendo-licensing": "^1.3.5",
     "@progress/kendo-svg-icons": "^2.0.0",
     "@progress/kendo-theme-default": "^7.2.0",
     "bootstrap": "^5.3.2",
@@ -48,7 +48,7 @@
   "devDependencies": {
     "@angular-devkit/build-angular": "^17.3.0",
     "@angular/cli": "^17.3.0",
-    "@angular/compiler-cli": "^17.2.4",
+    "@angular/compiler-cli": "^17.3.0",
     "@types/express": "^4.17.17",
     "@types/jasmine": "~5.1.0",
     "@types/node": "^18.18.0",
