@@ -5,7 +5,7 @@ import { HomeComponent } from './components/home.component';
 import { ErrorComponent } from './components/error.component';
 import { ProductsListComponent } from './components/products-list.component';
 import { GLOBAL} from './services/global';
-import { HttpClientModule,provideHttpClient } from '@angular/common/http';
+import { HttpClientModule,provideHttpClient, withFetch } from '@angular/common/http';
 import { ProductDetailsComponent } from './components/product-details.component';
 import { ProductEditComponent } from './components/product-edit.component';
 import { DateInputsModule } from '@progress/kendo-angular-dateinputs';
@@ -16,11 +16,7 @@ import { IndicatorsModule } from '@progress/kendo-angular-indicators';
 import {PanelBarModule } from '@progress/kendo-angular-layout';
 import { cartIcon, userIcon, SVGIcon } from '@progress/kendo-svg-icons';
 import { ButtonModule } from '@progress/kendo-angular-buttons';
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { BrowserModule } from '@angular/platform-browser';
-
-
-
+import { HttpClient } from '@angular/common/http';
 
 
 
